Bearer strategy checks for app secret proof
Account management front-end in Angular needs to be split into separate
ton-account application so that requests sent to the API will be signed
by an approved server. The only UI of ton-hub will be the authentication
dialogs. Until then, all requests sent to ton-hub which don't have app
secret proof will be rejected as unauthenticated.